Init network as package
